requirements.txt
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
robosuite==1.4.1
mujoco==2.3.7
gym==0.26.2
PyYAML==6.0.1
tqdm==4.66.4
tensorboard==2.16.2
wandb==0.16.6
scipy==1.12.0
opencv-python==4.9.0.80
